+++ client/CHANGELOG.md
@@ -6,6 +6,8 @@ All notable changes to this project will be documented in this file.
 
 ### Added
 
+- Added search input validation to ensure a user can't submit more than 100 words ([#847](https://github.com/project-lux/lux-frontend/issues/847)).
+
 ### Changed
 
 - Changed capitalization in advanced search for consistency ([#838](https://github.com/project-lux/lux-frontend/issues/838)).